[#234] Move get column stack bar options into util
