+++ frontend/src/pages/cases/visualizations/ChartCurrentFeasible.js
@@ -1,16 +1,6 @@
 import React, { useMemo } from "react";
 import Chart from "../../../components/chart";
-import {
-  incomeTargetChartOption,
-  Legend,
-  Color,
-  TextStyle,
-  thousandFormatter,
-  AxisShortLabelFormatter,
-  backgroundColor,
-  Easing,
-  LabelStyle,
-} from "../../../components/chart/options/common";
+import { getColumnStackBarOptions } from ".";
 
 const seriesTmp = [
   {
@@ -73,117 +63,6 @@ const seriesTmp = [
   },
 ];
 
-const getOptions = ({
-  xAxis = { name: "", axisLabel: {} },
-  yAxis = { name: "", min: 0, max: 0 },
-  origin = [],
-  series = [],
-  showLabel = false,
-}) => {
-  const legends = series.map((x) => x.name);
-  const xAxisData = origin.map((x) => x.name);
-
-  const options = {
-    legend: {
-      ...Legend,
-      data: legends,
-      top: 15,
-      left: "right",
-      orient: "vertical",
-    },
-    tooltip: {
-      trigger: "axis",
-      axisPointer: {
-        type: "shadow",
-      },
-      formatter: function (params) {
-        let res = "<div>";
-        res += "<b>" + params[0].axisValueLabel + "</b>";
-        res += "<ul style='list-style-type: none; margin: 0; padding: 0;'>";
-        params.forEach((param) => {
-          res += "<li>";
-          res += "<span>";
-          res += param.marker;
-          res += param.seriesName;
-          res += "</span>";
-          res +=
-            "<b style='float: right; margin-left: 12px;'>" +
-            thousandFormatter(param.value) +
-            "</b>";
-          res += "</li>";
-        });
-        res += "</ul>";
-        res += "</div>";
-        return res;
-      },
-      backgroundColor: "#ffffff",
-      ...TextStyle,
-    },
-    grid: {
-      top: 25,
-      left: 50,
-      right: 190,
-      bottom: 25,
-      show: true,
-      containLabel: true,
-      label: {
-        color: "#222",
-        ...TextStyle,
-      },
-    },
-    xAxis: {
-      ...xAxis,
-      nameTextStyle: { ...TextStyle },
-      nameLocation: "middle",
-      nameGap: 50,
-      boundaryGap: true,
-      type: "category",
-      data: xAxisData,
-      axisLabel: {
-        width: 100,
-        interval: 0,
-        overflow: "break",
-        ...TextStyle,
-        color: "#4b4b4e",
-        formatter: AxisShortLabelFormatter?.formatter,
-        ...xAxis.axisLabel,
-      },
-      axisTick: {
-        alignWithLabel: true,
-      },
-    },
-    yAxis: {
-      ...yAxis,
-      type: "value",
-      nameTextStyle: { ...TextStyle },
-      nameLocation: "middle",
-      nameGap: 75,
-      axisLabel: {
-        formatter: (e) => thousandFormatter(e),
-        ...TextStyle,
-        color: "#9292ab",
-      },
-    },
-    series: series.map((s) => {
-      if (s.type === "line") {
-        return s;
-      }
-      return {
-        ...s,
-        label: {
-          ...LabelStyle.label,
-          show: showLabel,
-          position: "right",
-        },
-      };
-    }),
-    ...Color,
-    ...backgroundColor,
-    ...Easing,
-  };
-  return options;
-};
-
 const ChartCurrentFeasible = ({
   dashboardData = [],
   currentCase,
@@ -197,20 +76,11 @@ const ChartCurrentFeasible = ({
           value: d?.[tmp.key] ? Math.round(d[tmp.key]) : 0,
         };
       });
-      let res = {
+      return {
         ...tmp,
-        emphasis: {
-          focus: "series",
-        },
+
         data: data,
       };
-      if (tmp.type === "line") {
-        res = {
-          ...incomeTargetChartOption,
-          ...res,
-        };
-      }
-      return res;
     });
   }, [dashboardData]);
 
@@ -219,7 +89,7 @@ const ChartCurrentFeasible = ({
       wrapper={false}
       type="BAR"
       loading={!chartData.length}
-      override={getOptions({
+      override={getColumnStackBarOptions({
         series: chartData,
         origin: dashboardData,
         yAxis: { name: `Income (${currentCase.currency})` },
